Create README.md
###Basic Python Project: TV Shows Ratings and Votes Analysis

What's This Project About?
Have you ever wondered if the most talked-about TV shows are also the highest-rated? That's what we're digging into here! This project is like a time machine back to the "Golden Age" of television, which started with "The Sopranos" in 1999. I'm looking at TV shows from that era to see if the ones that get tons of votes on IMDb are also the ones that score big with audiences.

The Data
I've got a file called /datasets/movies_and_shows.csv and it's packed with info on movies and shows. Here's what each column tells us:

name: The full name of the actor or director.
character: Who the actor played.
role: Was this person acting or directing?
title: What's the show or movie called?
type: Is it a show or a movie?
genres: What kind of show or movie is it?
release_year: When did it come out?
imdb_score: What did IMDb rate it?
imdb_votes: How many votes did it get on IMDb?
How to Get Through This Project
You'll see a Jupyter Notebook template I made. It's got places for you to write code and make notes about what you find. Here's the plan:

Part 1: Check Out the Data
Open up the data and see what's inside. Make some notes about what you find right there in the notebook.

Part 2: Clean Up the Data
Sometimes data can be a bit messy. I'll tidy up things like column names and take care of any missing stuff or repeats. There'll be spots in the notebook to jot down what you did.

Part 3: The Real Deal Analysis
This is where it gets juicy. I'll run some code to see if those popular shows are really the crowd-pleasers we think they are. And then, in the "Conclusion" bit of the notebook, I'll chat about what it all means.

Think of this project like a detective case. You've got a hunch, and now you're looking for clues to see if you're right. It's the same thing data pros do all the time in their jobs. Ready to play detective? Let's dive in!